refactor(selectors/global): extract some heavy imports to separate files [YTFRONT-3814]
