+++ packages/ui/src/ui/containers/SettingsPanel/settings-description.tsx
@@ -33,17 +33,17 @@ import ypath from '../../common/thor/ypath';
 import {AGGREGATOR_RADIO_ITEMS} from '../../constants/operations/statistics';
 import {NAMESPACES, SettingName} from '../../../shared/constants/settings';
 import {getRecentPagesInfo} from '../../store/selectors/slideoutMenu';
+import {getCurrentClusterNS} from '../../store/selectors/settings-ts';
 import SettingsMenuItem from '../../containers/SettingsMenu/SettingsMenuItem';
 import SettingsMenuRadio from '../../containers/SettingsMenu/SettingsMenuRadio';
 import SettingsMenuInput from '../SettingsMenu/SettingsMenuInput';
 import {
-    getCurrentClusterNS,
     getCurrentUserName,
     getGlobalMasterVersion,
     getGlobalSchedulerVersion,
     getHttpProxyVersion,
-    isDeveloperOrWatchMen,
 } from '../../store/selectors/global';
+import {isDeveloperOrWatchMen} from '../../store/selectors/global/is-developer';
 import {
     cellSizeRadioButtonItems,
     pageSizeRadioButtonItems,
