+++ packages/ui/src/ui/store/selectors/global/username.ts
@@ -1,0 +1,3 @@
+import {RootState} from '../../../store/reducers';
+
+export const getCurrentUserName = (state: RootState): string => state?.global?.login;
