+++ packages/ui/src/ui/store/selectors/slideoutMenu.ts
@@ -11,7 +11,9 @@ import {ClusterConfig} from '../../../shared/yt-types';
 import {RootState} from '../reducers';
 import {PAGE_ICONS_BY_ID} from '../../constants/slideoutMenu';
 import {getSettingsPagesOrder, getSettingsPagesPinned} from './settings-ts';
-import {getAllowedExperimentalPages, getClusterUiConfig, isDeveloper} from './global';
+import {getClusterUiConfig} from '../../store/selectors/global';
+import {isDeveloper} from '../../store/selectors/global/is-developer';
+import {getAllowedExperimentalPages} from './global/experimental-pages';
 import UIFactory from '../../UIFactory';
 import {Page} from '../../../shared/constants/settings';
